chore: Remove mobile token restrictions (#5026)
* chore: Remove mobile token restrictions

* Remove extra dependencies

* Update TicketTabNav_PurchaseTabScreen.tsx

* Update TicketTabNav_PurchaseTabScreen.tsx
