+++ src/select-travel-token-screen/SelectTravelTokenScreenComponent.tsx
@@ -9,25 +9,13 @@ import {
 } from '@atb/mobile-token';
 import {StyleSheet, Theme, useThemeContext} from '@atb/theme';
 import {ThemedTokenPhone, ThemedTokenTravelCard} from '@atb/theme/ThemedAssets';
-import {useFareContracts} from '@atb/ticketing';
-import {
-  dictionary,
-  getTextForLanguage,
-  TravelTokenTexts,
-  useTranslation,
-} from '@atb/translations';
+import {dictionary, TravelTokenTexts, useTranslation} from '@atb/translations';
 import {animateNextChange} from '@atb/utils/animation';
 import React, {useCallback, useEffect, useState} from 'react';
 import {ActivityIndicator, View} from 'react-native';
 import {ScrollView} from 'react-native-gesture-handler';
 import {RadioGroupSection} from '@atb/components/sections';
 import {useRemoteConfigContext} from '@atb/RemoteConfigContext';
-import {
-  findReferenceDataById,
-  isOfFareProductRef,
-  useFirestoreConfigurationContext,
-} from '@atb/configuration';
-import {useTimeContext} from '@atb/time';
 import {getDeviceNameWithUnitInfo} from './utils';
 import {TokenToggleInfo} from '@atb/token-toggle-info';
 import {useTokenToggleDetailsQuery} from '@atb/mobile-token/use-token-toggle-details';
@@ -37,20 +25,15 @@ type Props = {onAfterSave: () => void};
 
 export const SelectTravelTokenScreenComponent = ({onAfterSave}: Props) => {
   const styles = useStyles();
-  const {t, language} = useTranslation();
+  const {t} = useTranslation();
   const {theme} = useThemeContext();
 
   const {disable_travelcard} = useRemoteConfigContext();
-  const {fareProductTypeConfigs, preassignedFareProducts} =
-    useFirestoreConfigurationContext();
-
   const {completeOnboardingSection} = useOnboardingContext();
 
   const {tokens} = useMobileTokenContext();
   const toggleMutation = useToggleTokenMutation();
   const {data} = useTokenToggleDetailsQuery();
-
-  const {serverNow} = useTimeContext();
   const inspectableToken = tokens.find((t) => t.isInspectable);
 
   const [selectedType, setSelectedType] = useState<Token['type']>(
@@ -60,38 +43,6 @@ export const SelectTravelTokenScreenComponent = ({onAfterSave}: Props) => {
   const [selectedToken, setSelectedToken] = useState<Token | undefined>(
     inspectableToken,
   );
-
-  const {fareContracts: availableFareContracts} = useFareContracts(
-    {availability: 'available'},
-    serverNow,
-  );
-
-  const availableTravelRights = availableFareContracts.flatMap(
-    (fc) => fc.travelRights,
-  );
-
-  // Filter for unique travel rights config types
-  const availableFareContractsTypes = availableTravelRights.map(
-    (travelRight) => {
-      const preassignedFareProduct = findReferenceDataById(
-        preassignedFareProducts,
-        isOfFareProductRef(travelRight) ? travelRight.fareProductRef : '',
-      );
-
-      return (
-        preassignedFareProduct &&
-        fareProductTypeConfigs.find(
-          (c) => c.type === preassignedFareProduct.type,
-        )
-      );
-    },
-  );
-
-  const fareProductConfigWhichRequiresTokenOnMobile =
-    availableFareContractsTypes.find(
-      (fareProductTypeConfig) =>
-        fareProductTypeConfig?.configuration.requiresTokenOnMobile === true,
-    );
 
   useEffect(() => {
     // Whenever a user enters this screen, the onboarding is done.
@@ -121,14 +72,6 @@ export const SelectTravelTokenScreenComponent = ({onAfterSave}: Props) => {
 
   const travelCardToken = tokens?.find((t) => t.type === 'travel-card');
   const mobileTokens = tokens?.filter((t) => t.type === 'mobile');
-
-  // Shows an error message if switching to a t:card,
-  // but the current inspectable token is in the mobile AND
-  // requires mobile token
-  const requiresTokenOnMobile =
-    selectedType === 'travel-card' &&
-    inspectableToken?.type === 'mobile' &&
-    !!fareProductConfigWhichRequiresTokenOnMobile;
 
   return (
     <View style={styles.container}>
@@ -206,24 +149,6 @@ export const SelectTravelTokenScreenComponent = ({onAfterSave}: Props) => {
             isMarkdown={false}
           />
         )}
-        {requiresTokenOnMobile && (
-          <MessageInfoBox
-            type="error"
-            title={t(
-              TravelTokenTexts.toggleToken.notAllowedToUseTravelCardError.title,
-            )}
-            message={t(
-              TravelTokenTexts.toggleToken.notAllowedToUseTravelCardError.message(
-                getTextForLanguage(
-                  fareProductConfigWhichRequiresTokenOnMobile.name,
-                  language,
-                ) ?? '',
-              ),
-            )}
-            style={styles.errorMessageBox}
-            isMarkdown={false}
-          />
-        )}
         {selectedType === 'mobile' && mobileTokens?.length ? (
           <RadioGroupSection<Token>
             type="spacious"
@@ -254,16 +179,14 @@ export const SelectTravelTokenScreenComponent = ({onAfterSave}: Props) => {
         {toggleMutation.isLoading ? (
           <ActivityIndicator size="large" />
         ) : (
-          !requiresTokenOnMobile && (
-            <Button
-              expanded={true}
-              onPress={onSave}
-              text={t(TravelTokenTexts.toggleToken.saveButton)}
-              interactiveColor={theme.color.interactive[0]}
-              disabled={!selectedToken || (data?.toggleLimit ?? 0) < 1}
-              testID="confirmSelectionButton"
-            />
-          )
+          <Button
+            expanded={true}
+            onPress={onSave}
+            text={t(TravelTokenTexts.toggleToken.saveButton)}
+            interactiveColor={theme.color.interactive[0]}
+            disabled={!selectedToken || (data?.toggleLimit ?? 0) < 1}
+            testID="confirmSelectionButton"
+          />
         )}
       </ScrollView>
     </View>
