+++ src/stacks-hierarchy/Root_TabNavigatorStack/TabNav_TicketingStack/Ticketing_TicketTabNavStack/TicketTabNav_PurchaseTabScreen/TicketTabNav_PurchaseTabScreen.tsx
@@ -10,7 +10,6 @@ import {TicketTabNavScreenProps} from '../navigation-types';
 import {UpgradeSplash} from './Components/UpgradeSplash';
 import {FareProductTypeConfig} from '@atb/configuration';
 import {useAnalyticsContext} from '@atb/analytics';
-import {useMobileTokenContext} from '@atb/mobile-token';
 import {TariffZoneWithMetadata} from '@atb/tariff-zones-selector';
 import {StopPlaceFragment} from '@atb/api/types/generated/fragments/stop-places';
 import {TariffZone} from '@atb/configuration';
@@ -34,12 +33,6 @@ export const TicketTabNav_PurchaseTabScreen = ({navigation}: Props) => {
   const styles = useStyles();
   const analytics = useAnalyticsContext();
 
-  const {tokens, mobileTokenStatus} = useMobileTokenContext();
-  const inspectableToken = tokens.find((t) => t.isInspectable);
-  const hasInspectableMobileToken = inspectableToken?.type === 'mobile';
-  const hasMobileTokenError =
-    mobileTokenStatus === 'fallback' || mobileTokenStatus === 'error';
-
   if (must_upgrade_ticketing) return <UpgradeSplash />;
 
   const onProductSelect = (fareProductTypeConfig: FareProductTypeConfig) => {
@@ -51,41 +44,13 @@ export const TicketTabNav_PurchaseTabScreen = ({navigation}: Props) => {
       .forType(fareProductTypeConfig.type)
       .build();
 
-    if (authenticationType !== 'phone') {
-      if (
-        fareProductTypeConfig.configuration.requiresLogin &&
-        fareProductTypeConfig.configuration.requiresTokenOnMobile &&
-        !hasInspectableMobileToken
-      ) {
-        navigation.navigate('Root_LoginRequiredForFareProductScreen', {
-          selection,
-        });
-        return;
-      }
-
-      if (fareProductTypeConfig.configuration.requiresLogin) {
-        navigation.navigate('Root_LoginRequiredForFareProductScreen', {
-          selection,
-        });
-        return;
-      }
-    } else if (
-      fareProductTypeConfig.configuration.requiresTokenOnMobile &&
-      !hasInspectableMobileToken &&
-      !hasMobileTokenError
+    if (
+      authenticationType !== 'phone' &&
+      fareProductTypeConfig.configuration.requiresLogin
     ) {
-      navigation.navigate(
-        'Root_ActiveTokenOnPhoneRequiredForFareProductScreen',
-        {
-          nextScreen: {
-            screen: 'Root_PurchaseOverviewScreen',
-            params: {
-              selection,
-              mode: 'Ticket',
-            },
-          },
-        },
-      );
+      navigation.navigate('Root_LoginRequiredForFareProductScreen', {
+        selection,
+      });
       return;
     }
